requirements.txt
djangorestframework==3.2.4
django-rest-swagger==0.3.4